Moved events to preprocess.
