+++ TODO.txt
@@ -1,1 +1,1 @@
-# Remove headline # -*- coding: utf-8 -*-------------- recorder -------------    _recorder.py -- DONE    stream_recorder.py -- DONE        -> Test recorder        -> Rename amp_name to stream_name when instantiating the StreamReceiver.        -> sr.acquire() fills the buffer?        -> Check function pcl2fif when it will be updated.-------------- player --------------    stream_player.py -- DONE------------- triggers -------------triggers.trigger_def -- DONEtriggers.trigger -- TO CHECK---------- utils ----------utils.benchmark -> TODOutils.events    .brainvision.fix_default_event_values -- DONE    .change_event_values -- DONE    .find_event_channel -- DONE    .merge_events -> TODOutils.io    .convert2fif -> TODO & rename to read_raw.py    .export -- DONE    .io_file_dir -- DONE    .load_config -- DONE    .read_raw_fif -- DONEutils.layouts    .caps -- DONEutils.lsl    .lsl -- DONEutils.math    .math -- DONEutis.preprocess    preproces -> TODO -> To remove    rename_channels -- DONE    set_channel_types -- DONE    set_eeg_reference -- DONE    resample -- DONE    filter -- DONE    current_source_density -- DONEutils.timer    .timer -- DONE------Check all __init__.py from utils.# Set headline #!/usr/bin/env python3 for every executable scripts
+# Remove headline # -*- coding: utf-8 -*-------------- recorder -------------    _recorder.py -- DONE    stream_recorder.py -- DONE        -> Test recorder        -> Rename amp_name to stream_name when instantiating the StreamReceiver.        -> sr.acquire() fills the buffer?        -> Check function pcl2fif when it will be updated.-------------- player --------------    stream_player.py -- DONE------------- triggers -------------triggers.trigger_def -- DONEtriggers.trigger -- TO CHECK---------- utils ----------utils.benchmark -> TODOutils.io    .convert2fif -> TODO & rename to read_raw.py    .export -- DONE    .io_file_dir -- DONE    .load_config -- DONE    .read_raw_fif -- DONEutils.layouts    .caps -- DONEutils.lsl    .lsl -- DONEutils.math    .math -- DONEutis.preprocess    preproces -> TODO -> To remove    rename_channels -- DONE    set_channel_types -- DONE    set_eeg_reference -- DONE    resample -- DONE    filter -- DONE    current_source_density -- DONE    .events        .brainvision.fix_default_event_values -- DONE        .change_event_values -- DONE        .find_event_channel -- DONE        .merge_events -> TODOutils.timer    .timer -- DONE------Check all __init__.py from utils.# Set headline #!/usr/bin/env python3 for every executable scripts
